Silenced a compiler warning about a type mismatch in AppendEnvironment

+++ win/tclWinInit.c
@@ -527,7 +527,7 @@ AppendEnvironment(
      */
     for (shortlib = (char *) &lib[strlen(lib)-1]; shortlib>lib ; shortlib--) {
 	if (*shortlib == '/') {
-	    if (shortlib - lib == strlen(lib) - 1) {
+	    if ((unsigned)(shortlib - lib) == strlen(lib) - 1) {
 		Tcl_Panic("last character in lib cannot be '/'");
 	    }
 	    shortlib++;
